Provide default status slot-value if only one value is valid
When creating midi files, midi-messages must be instantiated with the correct values for the
status slot.  When only one value is valid, it is convenient to provide the default such that
a novice user of MIDI doesn't need to remember which value to use.

TODO: find a mechanism to fill in the channel from a dynamically-scoped variable into this
      status.