Sensible default for :status in midi-message
If both status-min and status-max are the same, only one :status
initarg really makes sense, so provide it in the defclass.

If status-max and status-min cover a range equal to the number of
midi-channels (ie 15), combine status-min and status-max with
*midi-channel* to supply a default to :status.

This is tested in usage/testmidi.lisp where no explicit :status is
needed anymore.